Former-commit-id: ff9931c8f2297ec3ba4c690e417868ae0f79069a

+++ include/hpx/par.h
@@ -77,7 +77,7 @@ int hpx_count_range_call(hpx_action_t action,
 			 const hpx_addr_t addr,
 			 const size_t count,
 			 const size_t increment,
-			 const size_t bsize,
+			 const uint32_t bsize,
 			 const size_t arg_size,
 			 void *const arg);
